treewide: changes to handle random vertex creation
Changes were made to both core and ui in order to support generation of random
vertices. Also, some ui components were added to allow the user to actually
move the nodes around and check the routing table in a different ui component
for each vertex (when it is clicked/released).